requirements.txt
cloudant==2.15.0
Flask==2.1.1
Flask_Cors==3.0.10
joblib==1.2.0
paho_mqtt==1.6.1
pyserial==3.5
